pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memorytwin"
version = "0.1.0"
description = "Episodic Memory Agent for Software Development - Dual Escriba + Oráculo Architecture"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Memory Twin Team"}
]
keywords = ["ai", "memory", "rag", "mcp", "copilot", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",  # For OpenRouter and other compatible providers
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "tenacity>=8.0.0",
    "sqlalchemy>=2.0.0",
    "scikit-learn>=1.0.0",  # For DBSCAN clustering in consolidation
]

[project.optional-dependencies]
# Web interface (Oráculo UI)
ui = [
    "gradio>=4.0.0",
]

# Observability
observability = [
    "langfuse>=2.0.0",
]

# SQL database (for scalability)
sql = [
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
]

# Alternative LLM providers
anthropic = [
    "anthropic>=0.30.0",
]

# All features
all = [
    "memorytwin[ui,observability,sql,anthropic]",
]

# Development
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
# Main command
mt = "memorytwin.escriba.cli:main"

memorytwin = "memorytwin.escriba.cli:main"
memorytwin-oraculo = "memorytwin.oraculo.app:main"
memorytwin-mcp = "memorytwin.mcp_server.server:main"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"src/memorytwin/observability.py" = ["E402"]
"src/memorytwin/mcp_server/server.py" = ["E402"]
"src/memorytwin/escriba/cli.py" = ["E402", "E501"]

[tool.coverage.run]
source = ["src/memorytwin"]
omit = [
    "src/memorytwin/escriba/cli.py",
    "src/memorytwin/oraculo/app.py",
    "src/memorytwin/escriba/project_analyzer.py",
    "src/memorytwin/observability.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
fail_under = 60
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
